refactor(HandleAST): fix the returnValueAST that doesn't handle some special cases and move the handleFunctions and returnValueHandleFunction in HandleAST to prevent from circular import of returnValueAST
